# Improve identifier parsing