… samples/Esprima.Benchmark/CharacterOld.cs
… src/Esprima/Character.Generated.cs