Improves identifier parsing to match the spec even in the astral planes (> \UFFFF) based on Unicode 15.0.

Also, does a minor optimization in the handling of the BMP range (\u0000..\uFFFF):

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |ParseProgram|angular-1.2.5|8,474.6 μs|36.10 μs|4293.9 KB|
| **New** |	|	| **8,345.3 μs (-2%)** | **80.26 μs** | **4293.9 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|1,140.7 μs|8.73 μs|673.02 KB|
| **New** |	|	| **1,138.9 μs (0%)** | **17.30 μs** | **673.02 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|6,661.2 μs|44.92 μs|3744.63 KB|
| **New** |	|	| **6,648.2 μs (0%)** | **40.23 μs** | **3744.63 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|10,209.9 μs|19.24 μs|5762.71 KB|
| **New** |	|	| **10,034.3 μs (-2%)** | **60.22 μs** | **5762.71 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|5,427.3 μs|31.40 μs|3074.04 KB|
| **New** |	|	| **5,300.6 μs (-2%)** | **32.30 μs** | **3074.04 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|974.3 μs|6.75 μs|568.59 KB|
| **New** |	|	| **938.8 μs (-4%)** | **6.16 μs** | **568.59 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|4,804.8 μs|7.93 μs|2827.59 KB|
| **New** |	|	| **4,667.7 μs (-3%)** | **18.89 μs** | **2827.59 KB (0%)** |