@lahma 2023-05-06
New tag pushed and build/publish has started: https://github.com/sebastienros/esprima-dotnet/actions/runs/4901904090/jobs/8753432865