@lahma 2023-05-08
Thank you! RC2 is now showing on nuget.org.